split types.rs + added additional helpers
