+++ common/gateway-requests/src/types/mod.rs
@@ -1,0 +1,18 @@
+// Copyright 2020-2024 - Nym Technologies SA <contact@nymtech.net>
+// SPDX-License-Identifier: Apache-2.0
+
+pub mod binary_request;
+pub mod binary_response;
+pub mod error;
+mod helpers;
+pub mod registration_handshake_wrapper;
+pub mod text_request;
+pub mod text_response;
+
+// just to preserve existing imports
+pub use binary_request::*;
+pub use binary_response::*;
+pub use error::*;
+pub use registration_handshake_wrapper::*;
+pub use text_request::*;
+pub use text_response::*;
